194_リファクタリング
