+++ app/controllers/users_controller.rb
@@ -4,7 +4,7 @@ class UsersController < ApplicationController
   before_action :ensure_admin, except: [:show]
   before_action :ensure_current_user, only: [:show, :status]
   before_action :set_users, only: [:show, :edit, :update, :destroy]
-  before_action :set_date, only: [:index, :status, :show]
+  before_action :set_date, only: [:status]
   before_action :set_user_chart_data, only: [:show]
 
   def index
