+++ app/helpers/users_helper.rb
@@ -1,10 +1,4 @@
 module UsersHelper
-  def set_month
-    today = Date.current
-    @year = today.year
-    @month = today.month
-  end
-
   def set_date
     today = Date.current
     @year = today.year
